+++ Evergreen/Private/Get-GitHubRepoRelease.ps1
@@ -3,6 +3,8 @@ Function Get-GitHubRepoRelease {
         .SYNOPSIS
             Calls the GitHub Releases API passed via $Uri, validates the response and returns a formatted object
             Example: https://api.github.com/repos/PowerShell/PowerShell/releases/latest
+
+            TODO: update to optionally return just the version number
     #>
     [OutputType([System.Management.Automation.PSObject])]
     [CmdletBinding(SupportsShouldProcess = $False)]
@@ -73,7 +75,6 @@ Function Get-GitHubRepoRelease {
             Break
         }
         Else {
-
             # If it's not a 403, return the exception to the pipeline
             Throw $_
         }
@@ -90,7 +91,7 @@ Function Get-GitHubRepoRelease {
                 ReferenceObject  = $script:resourceStrings.Properties.GitHub
                 DifferenceObject = (Get-Member -InputObject $item -MemberType NoteProperty)
                 PassThru         = $True
-                ErrorAction      = $script:resourceStrings.Preferences.ErrorAction
+                ErrorAction      = "Continue"
             }
             $missingProperties = Compare-Object @params
 
@@ -138,7 +139,9 @@ Function Get-GitHubRepoRelease {
                             Size         = $asset.size
                             URI          = $asset.browser_download_url
                         }
-                        Write-Output -InputObject $PSObject
+                        If ($PSObject.Platform -eq "Windows") {
+                            Write-Output -InputObject $PSObject
+                        }
                     }
                     Else {
                         Write-Verbose -Message "$($MyInvocation.MyCommand): Skip: $($asset.browser_download_url)."