+++ Evergreen/Private/Invoke-SystemNetRequest.ps1
@@ -29,7 +29,7 @@ Function Invoke-SystemNetRequest {
     }
     catch [System.Exception] {
         Write-Warning -Message "$($MyInvocation.MyCommand): Error at URI: $Uri."
-        Write-Warning -Message "$($MyInvocation.MyCommand): Response: $($webResponse.StatusCode) - $($webResponse.StatusDescription)"
+        Write-Warning -Message "$($MyInvocation.MyCommand): Response: $($_)."
         Write-Warning -Message "$($MyInvocation.MyCommand): For troubleshooting steps see: $($script:resourceStrings.Uri.Info)."
         #Throw "$($MyInvocation.MyCommand): $($_.Exception.Message)."
     }