+++ Evergreen/Private/Resolve-SystemNetWebRequest.ps1
@@ -26,7 +26,7 @@ Function Resolve-SystemNetWebRequest {
     }
     catch {
         Write-Warning -Message "$($MyInvocation.MyCommand): Error at URI: $Uri."
-        Write-Warning -Message "$($MyInvocation.MyCommand): Response: $($webResponse.StatusCode) - $($webResponse.StatusDescription)"
+        Write-Warning -Message "$($MyInvocation.MyCommand): Response: $($_)."
         Write-Warning -Message "$($MyInvocation.MyCommand): For troubleshooting steps see: $($script:resourceStrings.Uri.Info)."
         #Throw "$($MyInvocation.MyCommand): $($_.Exception.Message)."
     }