+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## VERSION
 
 * Adds `BlueJ`, `Postman`, `TableauDesktop`. Thanks to [@adotcoop](https://github.com/adotcoop)
+* Adds `MicrosoftAzureFunctionsCoreTools`, `MasterPackager`
 * Updates `Get-EvergreenApp` to sort output on the `Version` property in descending order
 * Updates internal functions `Invoke-SystemNetRequest`, `Invoke-WebRequestWrapper`, `Resolve-InvokeWebRequest`, `Resolve-SystemNetWebRequest` to now throw in the event of a source URL being unavailable, to allow for more graceful handling of vendor sources being temporarily unavailable